requirements.txt
django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
celery==5.4.0
redis==5.0.4
django-cors-headers==4.3.1